clk: move meson clk-regmap implementation to common code
I like this one better than qualcomms and wish to use it for the
PolarFire SoC clock drivers.

Acked-by: Neil Armstrong <neil.armstrong@linaro.org>
